+++ docs/Changelog.md
@@ -4,10 +4,14 @@ This page is a brief overview of each version.
 
 ---
 ## Unreleased
-- Expose searchCombinedEmbeddings function to AICore library for templates to use
 - Use a separate queue for indexing embeddings and summaries, to prevent blocking the main SB indexing thread
 - Refactor to use JSR for most Silverbullet imports, and lots of related changes
 - Reduced bundle size
+
+---
+## 0.3.2
+- Expose searchCombinedEmbeddings function to AICore library for templates to use
+- Add [[Providers/Ollama]] text/llm provider as a wrapper around openai provider
 ---
 ## 0.3.1
 - Set searchEmbeddings to false by default